requirements.txt
ccxt>=4.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
# pyfolio has compatibility issues with Python 3.10+
# Use alternative: quantstats or implement custom portfolio analysis
